Move get_nodes_between into mod utils
