+++ packages/sycamore-web/src/iter.rs
@@ -117,7 +117,7 @@ where
                     Box::new(move || {
                         // Get all nodes between start and end and reconcile with new nodes.
                         let mut new = flattened.get_clone();
-                        let mut old = get_nodes_between(&start_node, &end_node);
+                        let mut old = utils::get_nodes_between(&start_node, &end_node);
                         // We must include the end node in case `old` is empty (precondition for
                         // reconcile_fragments).
                         new.push(end_node.clone());
@@ -217,7 +217,7 @@ where
                     Box::new(move || {
                         // Get all nodes between start and end and reconcile with new nodes.
                         let mut new = flattened.get_clone();
-                        let mut old = get_nodes_between(&start_node, &end_node);
+                        let mut old = utils::get_nodes_between(&start_node, &end_node);
                         // We must include the end node in case `old` is empty (precondition for
                         // reconcile_fragments).
                         new.push(end_node.clone());
@@ -429,34 +429,4 @@ fn reconcile_fragments(parent: &web_sys::Node, a: &mut [web_sys::Node], b: &[web
             }
         }
     }
-}
-
-/// Get all nodes between `start` and `end`.
-///
-/// If `end` is before `start`, all nodes after `start` will be returned.
-///
-/// The range is exclusive so `start` and `end` will not be included.
-#[must_use]
-pub(crate) fn get_nodes_between(start: &web_sys::Node, end: &web_sys::Node) -> Vec<web_sys::Node> {
-    let parent = start.parent_node().unwrap();
-    debug_assert_eq!(
-        parent,
-        end.parent_node().unwrap(),
-        "parents of `start` and `end` do not match"
-    );
-
-    let mut nodes = Vec::new();
-
-    let mut next = start.next_sibling();
-    while let Some(current) = next {
-        let tmp = current.next_sibling();
-        if &current == end {
-            break;
-        } else {
-            nodes.push(current);
-        }
-        next = tmp;
-    }
-
-    nodes
 }
